+++ src/main/java/org/example/fibonacci/FibonacciRecursiveMatrixExponentiation.java
@@ -6,7 +6,7 @@ import static org.example.number.Constants.MATH_CONTEXT;
 
 public class FibonacciRecursiveMatrixExponentiation implements FibonacciStrategy<BigDecimal> {
     @Override
-    public BigDecimal solveFibonacci(int n) {
+    public BigDecimal solveFibonacci(long n) {
         if (n <= 1) {
             return BigDecimal.valueOf(n);
         }
@@ -23,7 +23,7 @@ public class FibonacciRecursiveMatrixExponentiation implements FibonacciStrategy
     }
 
     // reduces the problem size by half in each recursive call, time comp - O(log n)
-    private void matrixPower(BigDecimal[][] F, int n) {
+    private void matrixPower(BigDecimal[][] F, long n) {
         if (n == 1 || n == 0) return;
 
         // space comp - O(1)
