+++ src/main/java/org/example/input/NumberSource.java
@@ -1,5 +1,5 @@
 package org.example.input;
 
 public interface NumberSource {
-    int getNumber();
+    long getNumber();
 }
